Issue #80: Move workflow out from apps folder
