+++ devops/src/ax/devops/workflow/adc
@@ -6,6 +6,6 @@ monkey.patch_all()
 from ax.util.az_patch import az_patch
 az_patch()
 
-from ax.devops.apps.workflow.adc_entry import main
+from ax.devops.workflow.adc_entry import main
 if __name__ == '__main__':
     main()
