+++ devops/src/ax/devops/workflow/ax_workflow.py
@@ -9,7 +9,6 @@ Module for AXWorkflow
 """
 
 import json
-import time
 import uuid
 import time
