+++ devops/test/ax/test/devops/e2e/executor/conftest.py
@@ -3,6 +3,6 @@ import pytest
 
 @pytest.fixture()
 def workflow_executor():
-    from ax.devops.apps.workflow.ax_workflow_executor import AXWorkflowExecutor
+    from ax.devops.workflow.ax_workflow_executor import AXWorkflowExecutor
     workflowexecutor = AXWorkflowExecutor("test_workflow_id", "test_name1", "URL")
     yield workflowexecutor
